+++ src/Geopilot.Frontend/src/auth/AdminTemplate.jsx
@@ -1,7 +1,0 @@
-import { useAuth } from ".";
-
-export const AdminTemplate = ({ children }) => {
-  const { user } = useAuth();
-
-  return user?.isAdmin ? children : null;
-};