+++ src/Geopilot.Frontend/src/auth/LoggedInTemplate.jsx
@@ -1,7 +1,0 @@
-import { useAuth } from ".";
-
-export const LoggedInTemplate = ({ children }) => {
-  const { user } = useAuth();
-
-  return user ? children : null;
-};