+++ src/Geopilot.Frontend/src/auth/LoggedInTemplate.tsx
@@ -1,0 +1,12 @@
+import { useAuth } from ".";
+import { FC, ReactNode } from "react";
+
+interface LoggedInTemplateProps {
+  children: ReactNode;
+}
+
+export const LoggedInTemplate: FC<LoggedInTemplateProps> = ({ children }) => {
+  const { user } = useAuth();
+
+  return user ? children : null;
+};