+++ src/Geopilot.Frontend/src/auth/LoggedOutTemplate.jsx
@@ -1,7 +1,0 @@
-import { useAuth } from ".";
-
-export const LoggedOutTemplate = ({ children }) => {
-  const { user } = useAuth();
-
-  return user ? null : children;
-};