+++ src/Geopilot.Frontend/src/auth/LoggedOutTemplate.tsx
@@ -1,0 +1,12 @@
+import { useAuth } from ".";
+import { FC, ReactNode } from "react";
+
+interface LoggedOutTemplateProps {
+  children: ReactNode;
+}
+
+export const LoggedOutTemplate: FC<LoggedOutTemplateProps> = ({ children }) => {
+  const { user } = useAuth();
+
+  return user ? null : children;
+};